service-order/src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["MasaruFukazawa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
black = "^24.4.2"
isort = "^5.13.2"
bandit = "^1.7.8"
flake8 = "^7.0.0"
pydantic = "^2.7.1"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
requests = "^2.32.2"
cryptography = "^42.0.7"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
